requirements.txt
streamlit
openai
chromadb
python-dotenv
PyMuPDF
beautifulsoup4
pyyaml
